reeb-887
git-svn-id: http://svn.uhuila.cn/reeb/trunk@2691 96626eaa-6fa6-4472-8f5b-47c901fe3adc

+++ resale/home/app/models/job/resaler/YihaodianNewOrders.java
@@ -22,7 +22,7 @@ public class YihaodianNewOrders extends Job {
         //todo
         Resaler resaler = null;
         Goods goods = null;
-        long number = 1;
+        int number = 1;
         String mobile = "";
         String extOrderId= "";
         //todo log everything
@@ -50,7 +50,7 @@ public class YihaodianNewOrders extends Job {
         }
     }
 
-    public Order buildSeewiOrder(Resaler resaler, Goods goods, long number, String mobile) throws NotEnoughInventoryException{
+    public Order buildSeewiOrder(Resaler resaler, Goods goods, int number, String mobile) throws NotEnoughInventoryException{
         Order order = Order.createConsumeOrder(resaler.getId(), AccountType.RESALER);
         order.deliveryType = DeliveryType.SMS;
